Orders

+++ app/controllers/public/orders_controller.rb
@@ -34,25 +34,30 @@ class Public::OrdersController < ApplicationController
     @order = Order.new(order_params)
     @order.payment_way = params[:order][:payment_way].to_i
     @order.customer_id = current_customer.id
-    @order.save
-    if params[:order][:address_select] == "3"
-      shipping_address = ShippingAddress.new
-      shipping_address.postcode = @order.postcode
-      shipping_address.address = @order.address
-      shipping_address.name = @order.shipping_name
-      shipping_address.customer_id = current_customer.id
-      shipping_address.save
+    if @order.save
+      if params[:order][:address_select] == "3"
+        shipping_address = ShippingAddress.new
+        shipping_address.postcode = @order.postcode
+        shipping_address.address = @order.address
+        shipping_address.name = @order.shipping_name
+        shipping_address.customer_id = current_customer.id
+        shipping_address.save
+      end
+      cart_items.each do |cart|
+        order_details = OrderDetail.new
+        order_details.item_id = cart.item_id
+        order_details.purchase_price = cart.item.price
+        order_details.quantitiy = cart.quantitiy
+        order_details.order_id = @order.id
+        order_details.save
+      end
+      redirect_to orders_complete_path
+      cart_items.destroy_all
+    else
+      @order = Order.new
+      @shipping_addresses = current_customer.shipping_addresses
+      render :new
     end
-    cart_items.each do |cart|
-      order_details = OrderDetail.new
-      order_details.item_id = cart.item_id
-      order_details.purchase_price = cart.item.price
-      order_details.quantitiy = cart.quantitiy
-      order_details.order_id = @order.id
-      order_details.save
-    end
-    redirect_to orders_complete_path
-    cart_items.destroy_all
   end
 
   def show
